pass whole result of API call to PhotoList component
