+++ src/data/index.ts
@@ -3,14 +3,13 @@ import photoData from "./photo"
 
 function useApi() {
   const { albumQuery, albumsQuery } = albumData()
-  const { photosQuery, photoInAlbumQuery, photoQuery, photosQueryInfinite } = photoData()
+  const { photosQuery, photoInAlbumQuery, photoQuery } = photoData()
   return {
     albumQuery,
     albumsQuery,
     photosQuery,
     photoInAlbumQuery,
     photoQuery,
-    photosQueryInfinite,
   }
 }
