+++ src/pages/PhotoListPage.tsx
@@ -3,31 +3,14 @@ import PhotoList from "../components/PhotoList"
 import useApi from "../data"
 
 const PhotoListPage = () => {
-  const { photosQuery, photosQueryInfinite } = useApi()
+  const { photosQuery } = useApi()
 
-  const photos = photosQuery({ limit: 100 })
-  const infinitePhotos = photosQueryInfinite({
+  const photos = photosQuery({
     limit: 20,
     offset: 0,
   })
-  useEffect(() => {
-    function onScroll(event: any) {
-      const { scrollHeight, scrollTop, clientHeight } = event.target.scrollingElement
-      if (scrollHeight - scrollTop <= clientHeight * 1.5) {
-        if (infinitePhotos.hasNextPage) {
-          infinitePhotos.fetchNextPage()
-        }
-      }
-    }
 
-    document.addEventListener("scroll", onScroll)
-
-    return () => {
-      document.removeEventListener("scroll", onScroll)
-    }
-  }, [infinitePhotos])
-  //return <PhotoList {...photos} />
-  return <PhotoList {...infinitePhotos} />
+  return <PhotoList photoQuery={photos} />
 }
 
 export default PhotoListPage
